add a REVOLUTE_AXIS_ANY joint
I chose not to store the axis, so it is passed to the joint
initialization macros.

Also, to keep a common createJoint for all kinds of joints, I added
an optional axis argument there too. This is consistent with the way
URDF allows an axis tag, even for free flyers where it is meaning-less.

I'm not very fond of these choices, I'm open to changing them.