jac_point{,_chain}: do not reset the jacobian
setting the whole jacobian to zero at each cycle seems like a waste
of cpu: some blocks will always be zero and won't be touched by the
algorithms and the others will be overwritten anyway.

Given the user creates the jacobian, I'd say it is its responsability
to initialize it properly.

Moreover, depending on the offset and includeFreeFlyer parameters
I suspect we reset parts of the jacobian that the user might have carefully
initialized, forcing him to initialize them again at each cycle.