dllexport/import joint and body classes
Before this commit, only METAPOD_DLL{API,IMPORT,EXPORT} was defined.
However, there was several problems

* DLLEXPORT was used in place of DLLAPI

* Joint and Body classes were not exported

* we do not create any metapod DLL but instead simple_arm DLL and
  simple_humanoid DLL. So we need to pass the library name to macros
  which define DLL-exported classes.

This commit adds DLL API support for these two libs (and also others for
the models we generate for the benchmarks).

This commit also moves simple_arm and simple_humanoid to separate
directories.

Before, their cmake code was split between the project root (for files listing)
and the tests directory (for actual library creation). And a .cc file was
under include/.
Moving the cmake code was necessary in order to generate the libraries
config.hh (since the directory acts as a namespace).
The library names were also changed to add a metapod prefix. And the - in the
header directory was replaced with a _.

Eventually, this commit adds an helper function to generate config.hh.
One cannot use the one provided by JRL-CMAKE as it can only be called from the
project root directory.