add robotbuilder: a lib to generate robot models
and use it to generate the benchmarks

robotbuilder
------------

Enable a client program to generate the C++ source code for a metapod model.

One should:

 * call the setters configure the robotbuilder
 * call addLink repetively to describe its robot
 * call write to trigger the write.

The eventual files are generated from template texts.

In order to avoid the plain of manually embedding (and escaping) the
template text, the embedfile utility was used.

The library is not as clean as it could, but since the generated code
is changing a lot in the fusion branch, the robotbuilder cleaning is
done in this branch too.

The generated headers include a config.hh file generated from the same
template as jrl-cmake, except for the LIBRARY_VERSION macro, which has
been removed.

The resulting source code is independent from [jrl-]cmake.

embedfile
---------

Convert anu file into a C source file so that it can be included into an
application or a library.