Specialize mulMatrixTranposeBy because profiling shows it is costly.
The specialization is a test to see if the specialized operator
of the rotation matrices will provide the same performance,
than the specialization in the current commit.

The problem is that currently the RotationMatrixAboutX is not
used in rnea and crba. Indeed the jointmacros force sXp to
be a TransformT<RotationMatrix>.

Further optimization will call for a new mechanism to instanciate
Joints templated by rotation matrix for Xt.