fix no-op version of METAPOD_STATIC_ASSERT
The no-op implementation of METAPOD_STATIC_ASSERT did not support
a trailing semicolon. This commit hopefuly fixes it using a typefdef.

The include/metapod/algos/jac_point_chain.hh was lacking a semicolon
which made it fail to build when using a version different from the
no-op one. The semicolon is restored.

This commit also removes the last boost_static_assert.