This commit fixes a bug when using boost 1.48 for timers.
Martin Felis pointed out that an outer loop of one, with an inner
loop of 100000 was not equivalent with the benchmarks, to
an outer loop of 100 and an inner loop of 1000.

He also found out that it is related to a bug with timers.
This commit fix it, and closes #74.